# backend/docs/swagger.yaml
basePath: /
definitions:
...
  internal_handler.errorResponse:
    properties:
      code:
        $ref: '#/definitions/perchly-backend_internal_apierror.Code'
      error:
        type: string
    type: object
...
  internal_handler.personaTraitsResponse:
    properties:
      is_customized:
        type: boolean
      traits:
        $ref: '#/definitions/perchly-backend_internal_model.PersonaTraits'
    type: object
...
  internal_handler.setPersonaTraitsRequest:
    properties:
      directness:
        type: integer
      energy:
        type: integer
      humor:
        type: integer
      warmth:
        type: integer
      wisdom:
        type: integer
    type: object
...
  perchly-backend_internal_apierror.Code:
    enum:
    - invalid_request_body
    - unauthorized
    - internal_error
    - session_create_failed
    - invalid_email
    - email_code_send_failed
    - too_many_attempts
    - verification_code_expired
    - invalid_verification_code
    - verification_failed
    - invalid_display_name
    - user_not_found
    - profile_update_failed
    - invalid_refresh_token
    - session_refresh_failed
    - logout_failed
    - invalid_persona_id
    - persona_not_found
    - persona_fetch_failed
    - personas_list_failed
    - persona_not_age_appropriate
    - persona_trait_out_of_range
    - persona_traits_failed
    - invalid_conversation_id
    - conversation_not_found
    - conversation_forbidden
    - conversation_create_failed
    - conversations_list_failed
    - conversation_fetch_failed
    - conversation_context_missing
    - messages_list_failed
    - invalid_message_id
    - message_not_found
    - message_content_required
    - message_send_failed
    - streaming_unsupported
    - reaction_emoji_required
    - invalid_reaction_emoji
    - cannot_react_to_own_message
    - reaction_failed
    - invalid_age_range
    - invalid_mood_preference
    - invalid_selected_persona_id
    - preferred_name_required
    - preferred_name_invalid
    - onboarding_save_failed
    - onboarding_fetch_failed
    - onboarding_profile_not_found
    - quota_check_failed
    - quota_exceeded
    - chat_energy_failed
    type: string
    x-enum-varnames:
    - CodeInvalidRequestBody
    - CodeUnauthorized
    - CodeInternal
    - CodeSessionCreateFailed
    - CodeInvalidEmail
    - CodeEmailCodeSendFailed
    - CodeTooManyAttempts
    - CodeVerificationExpired
    - CodeInvalidCode
    - CodeVerificationFailed
    - CodeInvalidDisplayName
    - CodeUserNotFound
    - CodeProfileUpdateFailed
    - CodeInvalidRefreshToken
    - CodeSessionRefreshFailed
    - CodeLogoutFailed
    - CodeInvalidPersonaID
    - CodePersonaNotFound
    - CodePersonaFetchFailed
    - CodePersonasListFailed
    - CodePersonaNotAgeAppropriate
    - CodePersonaTraitOutOfRange
    - CodePersonaTraitsFailed
    - CodeInvalidConversationID
    - CodeConversationNotFound
    - CodeConversationForbidden
    - CodeConversationCreateFailed
    - CodeConversationsListFailed
    - CodeConversationFetchFailed
    - CodeConversationContextMissing
    - CodeMessagesListFailed
    - CodeInvalidMessageID
    - CodeMessageNotFound
    - CodeMessageContentRequired
    - CodeMessageSendFailed
    - CodeStreamingUnsupported
    - CodeReactionEmojiRequired
    - CodeInvalidReactionEmoji
    - CodeCannotReactToOwnMessage
    - CodeReactionFailed
    - CodeInvalidAgeRange
    - CodeInvalidMoodPreference
    - CodeInvalidSelectedPersonaID
    - CodePreferredNameRequired
    - CodePreferredNameInvalid
    - CodeOnboardingSaveFailed
    - CodeOnboardingFetchFailed
    - CodeOnboardingProfileNotFound
    - CodeQuotaCheckFailed
    - CodeQuotaExceeded
    - CodeChatEnergyFailed
...
  perchly-backend_internal_model.PersonaTraits:
    properties:
      directness:
        description: |-
          Directness: how blunt/unsoftened vs. gentle and hedged its
          responses are — this is the dial behind "harsher answers".
        type: integer
      energy:
        description: 'Energy: how energetic/enthusiastic vs. calm and measured its
          tone is.'
        type: integer
      humor:
        description: 'Humor: how often and readily it jokes vs. stays fully serious.'
        type: integer
      warmth:
        description: 'Warmth: how affectionate/close vs. distant and formal the persona
          is.'
        type: integer
      wisdom:
        description: 'Wisdom: how deep/reflective vs. surface-level its responses
          are.'
        type: integer
    type: object
info:
  contact: {}
  description: |-
    Perchly sohbet uygulamasının REST API'si — kimlik doğrulama, onboarding, persona, konuşma/mesajlaşma ve kota uçları.
    Kimlik doğrulama gerektiren uçlar için "Authorize" düğmesinden `Bearer <access_token>` girin.
  title: Perchly Backend API
  version: "1.0"
paths:
...
  /personas/{id}/traits:
    delete:
      description: Kullanıcının özelleştirmesini siler; persona varsayılan traits'ine
        geri döner.
      parameters:
      - description: Persona ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.personaTraitsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Kişilik dial'larını sıfırla
      tags:
      - persona-traits
    get:
      description: Kullanıcının bu persona için özelleştirdiği (veya henüz özelleştirmediyse
        persona varsayılanı) warmth/humor/wisdom/directness/energy değerleri.
      parameters:
      - description: Persona ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.personaTraitsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Kişilik dial'larını getir
      tags:
      - persona-traits
    put:
      consumes:
      - application/json
      description: Beş dial de her çağrıda zorunludur (kısmi patch değil, tam panel
        state'i gönderilir); 0-100 aralığı dışı 400 döner.
      parameters:
      - description: Persona ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: warmth, humor, wisdom, directness, energy (0-100)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.setPersonaTraitsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.personaTraitsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Kişilik dial'larını ayarla
      tags:
      - persona-traits
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"